fix: remove FileStatus  from FileData
- Fileクラスが大きくなりすぎるため、FileStatus情報を別クラスとして切り出す
